requirements.txt
git+https://github.com/wyapx/CAI
revChatGPT
PyYaml